Update challenge ComputeThePerimeterOfAPolygon, change entered point type from double to float

+++ Functional Programming/ComputeThePerimeterOfAPolygon/app/Main.hs
@@ -26,7 +26,7 @@ main = do
   numberOfPointsToBeEntered <- readLn :: IO Int -- Read and bind number of points to be entered
 
   points <- replicateM numberOfPointsToBeEntered $ do -- For each point to be entered...
-    (x : y : _) <- (read <$>) . words <$> getLine :: IO [Double] -- Read and bind the point to x and y
+    (x : y : _) <- (read <$>) . words <$> getLine :: IO [Float] -- Read and bind the point to x and y
     return (x, y) -- Return the point as (x, y)
   
   print $ perimeter points -- Print the perimeter of entered polygon (as a list of points)
